requirements.txt
gunicorn
requests
flask
yfinance
panda
pytrends
forex-python